backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
simple-salesforce==1.12.5
python-dotenv==1.0.0
groq==0.4.2
pandas==2.0.0
openpyxl==3.10.0
pydantic==2.0.0
requests==2.31.0